Initial commit.
Getting my current Klipper configuration files into git finally.

Had to create a special key attached to my github repo in order to make
sure that it had the appropriate security.

I have gitignore'd a lot of stuff, but those are files that deal with
other configurations that I am not likely going to be able to transfer.
Such as my webcam settings which I believe would probably need to be
done over again if I were to switch to a new printer.